requirements.txt
bcrypt==4.1.3
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
httptools==0.6.1
marshmallow==3.26.1
PyJWT==2.8.0
pymongo==4.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3